202512_avis_stores_doc_eleve/requirements.txt
streamlit
pandas
google-generativeai
google-play-scraper
requests
plotly
python-dotenv
pydantic
nbformat>=4.2.0
ipywidgets
